+++ .TODO
@@ -13,6 +13,7 @@
 * Сделать проверки на наличие токенов в куки и лс, чтобы не дублировать их и не вызвать зацикленность
 * Убрать лишние useMemo и useCallback для наилучшей оптимизации
 * Деструктурировать компонент BurgerConstructor. Слишком много строк для одного компонента
+* Сделать валидацию форм
 
 Какие компоненты нужно мемоизировать и переработать:
 * BurgerConstructor
