+++ src/components/BurgerConstructor/BurgerConstructor.module.scss
@@ -14,75 +14,11 @@
 	height: 100%;
 }
 
-.burgerConstructor {
-	display: flex;
-	flex-direction: column;
-	align-items: flex-end;
-	row-gap: 16px;
-	flex: 1;
 
-	width: 100%;
 
-	border-radius: 50px;
 
-	transition: box-shadow $trans-fast;
 
-	&--hover {
-		box-shadow: $glow-small;
-	}
-}
 
-.ingredients {
-	width: 100%;
-
-	flex-grow: 1;
-
-	position: relative;
-
-	&__plug {
-		width: 100%;
-
-		padding: 0 16px;
-
-		display: flex;
-		flex-direction: column;
-		align-items: flex-end;
-	}
-
-	&__container {
-		width: 100%;
-
-		position: absolute;
-		inset: 0;
-
-		display: flex;
-		flex-direction: column;
-
-		gap: 16px 0;
-
-		padding: 0 16px;
-
-		@include scrollbar;
-
-		border-radius: 50px;
-
-		transition: box-shadow $trans-fast;
-
-		&--onDrop {
-			box-shadow: $glow-small;
-		}
-	}
-}
-
-.blockedElement {
-	width: 100%;
-	padding: 0 16px;
-
-	display: flex;
-	flex-direction: row;
-	align-items: center;
-	justify-content: flex-end;
-}
 
 .choiceOffer {
 	width: 100%;
