+++ src/components/BurgerConstructor/ConstructorBody/ConstructorBody.module.scss
@@ -1,0 +1,69 @@
+//? === Constructor Body ===
+
+@import 'scss/vars';
+
+.wrapper {
+	display: flex;
+	flex-direction: column;
+	align-items: flex-end;
+	row-gap: 16px;
+	flex: 1;
+
+	width: 100%;
+
+	border-radius: 50px;
+
+	transition: box-shadow $trans-fast;
+
+	&_hover {
+		box-shadow: $glow-small;
+	}
+}
+
+.bun {
+	width: 100%;
+	padding: 0 16px;
+
+	display: flex;
+	flex-direction: row;
+	align-items: center;
+	justify-content: flex-end;
+}
+
+.ingredients {
+	width: 100%;
+
+	flex-grow: 1;
+
+	position: relative;
+
+	&__plug {
+		width: 100%;
+
+		padding: 0 16px;
+
+		display: flex;
+		flex-direction: column;
+		align-items: flex-end;
+	}
+
+	&__container {
+		width: 100%;
+
+		position: absolute;
+		inset: 0;
+
+		display: flex;
+		flex-direction: column;
+
+		gap: 16px 0;
+
+		padding: 0 16px;
+
+		@include scrollbar;
+
+		border-radius: 50px;
+
+		transition: box-shadow $trans-fast;
+	}
+}
