+++ src/components/ConstructorItem/ConstructorItem.module.scss
@@ -6,7 +6,7 @@
 	align-items: center;
 	gap: 8px;
 
-	&--draggable {
+	&_draggable {
 		opacity: 0;
 	}
 }
