+++ src/components/Modal/Modal.module.scss
@@ -29,7 +29,7 @@
 	border-radius: 40px;
 	background-color: $bg-modal;
 
-	&--withHeading {
+	&_withHeading {
 		padding-bottom: 60px;
 
 		.header {
